Revert "WA chat bot"
This reverts commit 252a5b0608d34fd4d484f1fc738cd3a72838dbd2.
