+++ apps/api/package.json
@@ -47,7 +47,6 @@
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.0",
     "uuid": "^11.1.0",
-    "x-hub-signature": "^2.1.2",
     "zod": "^3.20.6"
   },
   "devDependencies": {
