+++ apps/api/src/controllers/helper/uploadtos3.ts
@@ -237,7 +237,6 @@ export async function uploadFileToS3(
     type: string,
     data: Readable | Buffer,
     key: string,
-    mimeType?: string,
 ): Promise<string> {
     const bucketName = getBucketFromTypeAndFolderName(type, '').split('/')[0];
 
@@ -246,7 +245,6 @@ export async function uploadFileToS3(
         Key: key,
         Body: data,
         ACL: 'public-read',
-        ContentType: mimeType,
     };
 
     const uploadResult = await s3.upload(uploadParams).promise();
