+++ apps/api/src/server.ts
@@ -42,7 +42,6 @@ import { cleanUpGiftCardLiveTemplates, startAppV2ErrorLogsCronJob } from "./serv
 import utilsRoutes from "./routes/utilsRoutes";
 import emailTemplateRoutes from "./routes/emailTemplateRoutes";
 import paymentRoutes from "./routes/paymentRoutes";
-import whatsAppRoutes from "./routes/whatsAppRoutes";
 import viewRoutes from "./routes/viewRoutes";
 import swaggerJSDoc from "swagger-jsdoc";
 
@@ -127,7 +126,6 @@ const initExpressApp = (app: express.Application) => {
   app.use("/api/utils", utilsRoutes );
   app.use("/api/email-templates", emailTemplateRoutes );
   app.use("/api/payments", paymentRoutes );
-  app.use("/api/wa", whatsAppRoutes);
   app.use("/api/view", viewRoutes);
 
   app.use('/api/doc', swaggerUi.serve, swaggerSpecification())
