--- apps/api/src/services/WhatsApp/logResponseError.ts
@@ -1,11 +1,0 @@
-
-function logResponseError(error: any) {
-    if (error?.response?.data) {
-        console.log(JSON.stringify(error.response.data, null, 2));
-    }
-    else {
-        console.log(error);
-    }
-}
-
-export { logResponseError }
