--- apps/api/src/services/WhatsApp/messageStatuses.ts
@@ -1,10 +1,0 @@
-
-const messageStatuses = {
-  read : {
-    messaging_product: "whatsapp",
-    status: "read",
-    message_id: ""
-  }
-}
-
-export { messageStatuses }
